Rename all .h files to .hpp
