+++ ground_station/src/console.hpp
@@ -1,5 +1,5 @@
 /******************************************************************************
- * file    console.h
+ * file    console.hpp
  *******************************************************************************
  * brief   Wrapper for Serial Debugging (threadsafe, buffered, colorized, etc.)
  *******************************************************************************
