+++ ground_station/src/logging/recorder.cpp
@@ -1,5 +1,5 @@
 
-#include "recorder.h"
+#include "recorder.hpp"
 
 bool Recorder::begin() {
   int32_t number = 0;
