+++ ground_station/src/systemParser.hpp
@@ -1,5 +1,5 @@
 /******************************************************************************
- * file    systemParser.h
+ * file    systemParser.hpp
  *******************************************************************************
  * brief   JSON-File Parser for system configuration
  *******************************************************************************
