+++ ground_station/src/utils.hpp
@@ -1,5 +1,5 @@
 /******************************************************************************
- * file    utils.h
+ * file    utils.hpp
  *******************************************************************************
  * brief   General utilities for file system support, MSC, configuration, etc.
  *******************************************************************************
